refactor: move `resolveCodexHome` and `resolveGeminiCredsDir` into their provider modules
Keeps `cli-utils.ts` provider-neutral (only `seedAuthFile`, diagnostics,
and `STALE_TIMEOUT_MS`). Co-locates the resolver unit tests with each
provider module.

Addresses manki-review review comments on PR #703
https://github.com/manki-review/manki/pull/703#discussion_r3212345767
https://github.com/manki-review/manki/pull/703#discussion_r3212378284
