+++ src/providers/cli-utils.test.ts
@@ -2,7 +2,7 @@ import { existsSync, mkdtempSync, readFileSync, rmSync, statSync, writeFileSync 
 import { tmpdir } from 'os';
 import { join } from 'path';
 
-import { buildTimeoutDiagnostics, resolveCodexHome, resolveGeminiCredsDir, sanitizeLogOutput, seedAuthFile } from './cli-utils';
+import { buildTimeoutDiagnostics, sanitizeLogOutput, seedAuthFile } from './cli-utils';
 
 function encode(json: unknown): string {
   return Buffer.from(JSON.stringify(json), 'utf8').toString('base64');
@@ -176,59 +176,3 @@ describe('seedAuthFile', () => {
   });
 });
 
-describe('resolveCodexHome', () => {
-  let savedCodexHome: string | undefined;
-  let savedHome: string | undefined;
-
-  beforeEach(() => {
-    savedCodexHome = process.env.CODEX_HOME;
-    savedHome = process.env.HOME;
-  });
-
-  afterEach(() => {
-    if (savedCodexHome === undefined) delete process.env.CODEX_HOME;
-    else process.env.CODEX_HOME = savedCodexHome;
-    if (savedHome === undefined) delete process.env.HOME;
-    else process.env.HOME = savedHome;
-  });
-
-  it('returns CODEX_HOME when set', () => {
-    process.env.CODEX_HOME = '/explicit';
-    expect(resolveCodexHome()).toBe('/explicit');
-  });
-
-  it('returns $HOME/.codex when CODEX_HOME is unset', () => {
-    delete process.env.CODEX_HOME;
-    process.env.HOME = '/h';
-    expect(resolveCodexHome()).toBe('/h/.codex');
-  });
-
-  it('throws when neither CODEX_HOME nor HOME is set', () => {
-    delete process.env.CODEX_HOME;
-    delete process.env.HOME;
-    expect(() => resolveCodexHome()).toThrow(/neither \$CODEX_HOME nor \$HOME/);
-  });
-});
-
-describe('resolveGeminiCredsDir', () => {
-  let savedHome: string | undefined;
-
-  beforeEach(() => {
-    savedHome = process.env.HOME;
-  });
-
-  afterEach(() => {
-    if (savedHome === undefined) delete process.env.HOME;
-    else process.env.HOME = savedHome;
-  });
-
-  it('returns $HOME/.gemini when HOME is set', () => {
-    process.env.HOME = '/h';
-    expect(resolveGeminiCredsDir()).toBe('/h/.gemini');
-  });
-
-  it('throws when HOME is not set', () => {
-    delete process.env.HOME;
-    expect(() => resolveGeminiCredsDir()).toThrow(/\$HOME is not set/);
-  });
-});
